Value looper related fixes
Turned ValueLooper into abstract class and added abstract method
setupValueLooper(?)
Also made new class BlendModeValueLooper - which is a subclass of
ValueLooper<> and the BMVL implements the abstract setup method. This
way the initialization of the looper is done in this one place.
(partitioned).
in sun.pde - removed call to setupValueLooper() method
from setupValueLooper - actually removed the method for setting up the
looper and moved it to the BMVL.
Also the BMVL instance is inline created.  In future need to move this
variable to main tab.